@@ -1,7 +1,6 @@
 package com.ouyang.boot.quartz.boot整合quartz;
 
 import org.quartz.Scheduler;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.PropertiesFactoryBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,6 +8,7 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.scheduling.quartz.SchedulerFactoryBean;
 
+import javax.annotation.Resource;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.util.Properties;
@@ -21,40 +21,46 @@ import java.util.concurrent.Executor;
 @Configuration
 public class SchedulerConfig {
 
-    @Autowired
+    @Resource
     private DataSource dataSource;
 
     @Bean
     public Scheduler scheduler() throws IOException {
-       return schedulerFactoryBean().getScheduler();
+        return schedulerFactoryBean().getScheduler();
     }
+
     @Bean
     public SchedulerFactoryBean schedulerFactoryBean() throws IOException {
-        SchedulerFactoryBean factory = new SchedulerFactoryBean();
-        factory.setSchedulerName("cluster_schedule");
-        factory.setDataSource(dataSource);
-        factory.setApplicationContextSchedulerContextKey("application");
-        factory.setQuartzProperties(properties());
-        factory.setTaskExecutor(schedulerThreadPool());
-        factory.setStartupDelay(0);
-        return factory;
+        return new SchedulerFactoryBean(){{
+            setSchedulerName("cluster_schedule");
+            setDataSource(dataSource);
+            setApplicationContextSchedulerContextKey("application");
+            setQuartzProperties(properties());
+            setTaskExecutor(schedulerThreadPool());
+            setStartupDelay(0);
+        }};
+
+
     }
 
     @Bean
     public Properties properties() throws IOException {
-        PropertiesFactoryBean propertiesFactoryBean = new PropertiesFactoryBean();
-        propertiesFactoryBean.setLocation(new ClassPathResource(""));
-        propertiesFactoryBean.afterPropertiesSet();
-        return propertiesFactoryBean.getObject();
+        return new PropertiesFactoryBean() {{
+            setLocation(new ClassPathResource(""));
+            afterPropertiesSet();
+        }}.getObject();
+
+
     }
 
     @Bean
     public Executor schedulerThreadPool() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
         int processors = Runtime.getRuntime().availableProcessors();
-        executor.setCorePoolSize(processors);
-        executor.setMaxPoolSize(processors);
-        executor.setQueueCapacity(processors);
-        return executor;
+        return new ThreadPoolTaskExecutor() {{
+            setCorePoolSize(processors);
+            setMaxPoolSize(processors);
+            setQueueCapacity(processors);
+        }};
+
     }
 }
